themes: exposure: index: fix incorrect logic for no_big_gallery_cover
Big gallery covers should be used for lines where only one gallery
cover appears.

With the current logic, if there is a prime number of galleries (except
2 and 3), first one and all galleries whose index is prime (except 2nd
and 3rd) will have a big cover.

In the end, all it matters is that if the galleries_line contains only
one gallery, that gallery should have a big cover.

Signed-off-by: Quentin Schulz <foss@0leil.net>