prosopopee: do not get logs from third party modules
Loggers work by hierarchy. The parent always overrides whatever the
child logger has already defined. This applies to the loglevel, which is
changed in prosopopee according to the --log-level argument.

Since the root logger (gotten with logger = logging.getLogger()) is the
parent of ALL loggers which could be declared in any third party module,
prosopopee's loglevel also applies to those modules which is usually not
wanted especially when prosopopee's default loglevel is the highest
available.

This is very annoying with Pillow since it's pretty verbose when saving
files.

Instead, let's declare a logger for prosopopee only. Unfortunately,
since the package layout is unconventional (all *.py files in the same
directory, instead of subdirs), the recommended
logger = logging.getLogger(__name__) cannot be used because __name__ is
__main__ in prosopopee.py, and the filename of the file in which it is
used (e.g. in cache.py, it'll be cache). Which means they're not related
in the eyes of the logging module and prosopopee.py's loglevel will not
apply to other *.py files in the project.

Instead the expected value of __name__ for more conventional packaging
layouts is simulated by appending prosopopee. in front of __name__
except for prosopopee.py which is the parent logger and thus will be
simply named prosopopee.

Since prosopopee's logger is not the root logger anymore, NOTSET
loglevel cannot be used anymore because its meaning is basically
"offload messages to parent logger" and the root logger has a default
loglevel of WARNING, meaning prosopopee's default loglevel will not
print anything labelled as INFO or DEBUG.

c.f. https://stackoverflow.com/a/50755200

Signed-off-by: Quentin Schulz <foss@0leil.net>